is RPM int or float, part X fix #6481

+++ firmware/controllers/engine_cycle/rpm_calculator.h
@@ -112,7 +112,7 @@ public:
 	/**
 	 * this is RPM on previous engine cycle.
 	 */
-	int previousRpmValue = 0;
+	float previousRpmValue = 0;
 
 	/**
 	 * This is a performance optimization: let's pre-calculate this each time RPM changes
